+++ src/lib/Fund.js
@@ -1,23 +1,29 @@
 import moment from 'moment'
-import { calcFinancialYear } from './Utils'
+import { calcFinancialYear, currentFinancialYear } from './Utils'
 
 class Fund {
   fyStartDate
   fyEndDate
   formedDate
+  currentFY
 
   constructor(fyStartDate, fyEndDate, formedDate) {
     this.fyStartDate = fyStartDate
     this.fyEndDate = fyEndDate
     this.formedDate = formedDate
+    this.currentFY = currentFinancialYear
+  }
+
+  isEmpty() {
+    return this.fyStartDate === undefined && this.fyEndDate === undefined && this.formedDate === undefined
   }
 
   startDate() {
-    return this.fyStartDate || this.formedDate || moment('1970-01-01')
+    return this.fyStartDate || this.formedDate || moment(`${this.currentFY - 1}-07-01`)
   }
 
   endDate() {
-    return this.fyEndDate || moment()
+    return this.fyEndDate || moment(`${this.currentFY}-06-30`)
   }
 
   financialYear() {